+++ es6/tests/index.js
@@ -7,6 +7,7 @@ if (path.resolve) {
 setStartFunction(startTest);
 
 function startTest() {
+	require("./e2e/text.js");
 	require("./e2e/base.js");
 	require("./e2e/xml-templater.js");
 	require("./e2e/errors.js");